m1n1.fw.aop: Split off ASCArgumentSection from experiments/aop.py
