+++ proxyclient/m1n1/fw/aop/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: MIT
+from .bootargs import ASCArgumentSection
+
+__all__ = ["ASCArgumentSection"]
